+++ src/module/Process.hpp
@@ -18,5 +18,6 @@ namespace be {
         static JSValue gc(JSContext *ctx, JSValueConst this_val, int argc, JSValueConst *argv) ;
         static JSValue ref(JSContext *ctx, JSValueConst this_val, int argc, JSValueConst *argv) ;
         static JSValue getRunningTime(JSContext *ctx, JSValueConst this_val, int argc, JSValueConst *argv) ;
+        static JSValue getChipTemperature(JSContext *ctx, JSValueConst this_val, int argc, JSValueConst *argv) ;
     } ;
 }